remove dots from function factories
The dots are only effective when specified within a call to a factory-made function, not when calling the respective function factory itself. Both the presence of the dots in the factories and the attending documentation were misleading and have now been removed. Dots can still be used in factory-made functions.