conditionally convert helper's input to character
This fixes a bug in `modify_to_length()` which used to cause `audit_seq()` to fail when called on the output of `debit_map_seq()`. Apparently, the bug was caused by the input to the internal helper function `modify_to_length()` being logical. Logical input to this function, therefore, is now converted to character. Also, I added some comments to structure the code.