fix `unround()` bugs
These bugs occurred in the `rounding_bounds_scalar()` helper. They lead to wrong reconstructions of rounding bounds when the `rounding` parameter of `unround()` was specified as `"trunc"` or `"anti_trunc"`. Also, the new version uses `switch()` instead of calling `dplyr::case_when()` because `switch()` is much faster, and `rounding_bounds_scalar()` needs to be subsequently vectorized in either case; so it doesn't matter that `case_when()` is vectorized but `switch()` is not.