cut ref to `grim()` details for param inheritance
The documentation for `grimmer()` includes `@inheritParams grim`, which leads to `grimmer()`'s non-documented parameters to be inherited from `grim()` if parameters by the same names are documented there. However, one such parameter used to refer to `grim()`'s Details section, which does not make sense with `grimmer()`. This reference was now removed.