unguard proritizing operations, some polishing
The reason for the unguarding is that these prioritizations might be relevant even to factory-made functions that were created without specifying `.arg_list`, and within which `.arg_list` is therefore `NULL`.

Regarding "polishing", I renamed `.show_more_cols` to `.col_control` and did some minor refactoring as well as cleanup.