handwrite `write_doc_audit_seq()`
It used to be the product of a function factory called `write_doc_factory_map_audit_section()` which turned out not to be very useful apart from this one special case. This suggested refactoring `write_doc_audit_seq()` as a conventional, "handwritten" function. What tipped me over was the need to fix a bug in the way the function handles `key_args` inputs of length > 2. The fix will follow soon.